+++ src/config/default/interrupts.h
@@ -57,6 +57,7 @@
 // *****************************************************************************
 // *****************************************************************************
 void CORE_TIMER_InterruptHandler( void );
+void TIMER_3_InterruptHandler( void );
 void SPI_2_InterruptHandler( void );
 void UART_2_InterruptHandler( void );
 void NVM_InterruptHandler( void );