+++ src/config/default/peripheral/evic/plib_evic.c
@@ -56,6 +56,7 @@ void EVIC_Initialize( void )
 
     /* Set up priority and subpriority of enabled interrupts */
     IPC0SET = 0x4U | 0x0U;  /* CORE_TIMER:  Priority 1 / Subpriority 0 */
+    IPC3SET = 0x4U | 0x0U;  /* TIMER_3:  Priority 1 / Subpriority 0 */
     IPC7SET = 0x4000000U | 0x0U;  /* SPI_2:  Priority 1 / Subpriority 0 */
     IPC8SET = 0x4U | 0x0U;  /* UART_2:  Priority 1 / Subpriority 0 */
     IPC11SET = 0x4U | 0x0U;  /* FCE:  Priority 1 / Subpriority 0 */